// SGUEES-SPA/src/app/pages/SelectionHiring/sc-descriptor-responsabilidad-cargo/models/sc-descriptor-responsabilidad-cargo.ts
export interface ScDescriptorResponsabilidadCargo {
	CORR_EMPRESA: number;
	CORR_RESPONSABILIDAD: number;
	NOMBRE_RESPONSABILIDAD: string;
	ESTADO_RESPONSABILIDAD: boolean;
	USUARIO_CREA: string;
	ESTACION_CREA: string;
	FECHA_CREA: Date;
	USUARIO_ACTU: string;
	ESTACION_ACTU: string;
	FECHA_ACTU: Date;
}